+++ docs/src/Avatar.doc.js
@@ -45,6 +45,12 @@ card(
         type: 'boolean',
         defaultValue: false,
       },
+      {
+        name: 'icon',
+        type: '"check-circle" | "pinterest"',
+        defaultValue: 'check-circle',
+        description: 'This is the icon shown only when "verified=true"',
+      },
     ]}
   />
 );
@@ -124,4 +130,22 @@ card(
   />
 );
 
+card(
+  <Example
+    description={`
+    We can also changed \`verified\` icon to the Pinterest logo by setting the
+    \`icon\` prop to \`pinterest\`.
+  `}
+    name="Verified"
+    defaultCode={`
+<Avatar
+  name="Shanice"
+  size="lg"
+  src="${shanice}"
+  verified
+  icon="pinterest"
+/>
+  `}
+  />
+);
 export default cards;