+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Internal: Updated rollup build tools to use stable/predictable minified CSS classnames (#455)
 - Video: add children prop (#464)
 - Icon: add new icon reorder-images (#466)
+- Avatar: Add property to use pinterest icon for verified (#460)
 
 ### Patch